+++ Backend/src/controllers/user/user.controller.ts
@@ -1,13 +1,12 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, Req, UseGuards } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, ForbiddenException, Get, Param, Patch, Post, Req, UseGuards } from '@nestjs/common';
 import { UserService } from 'src/core/services';
-import { CreateUserReqApiDto } from './dto/create-user.dto';
 import { UpdateUserReqApiDto } from './dto/update-user.dto';
 import { JwtAuthGuard, ERole } from 'src/core/jwt-auth.guard';
 import { RolesGuard } from '../auth/roles.guard';
 import { RequiredRoles } from '../auth/roles.decorator';
 
 @Controller('user')
-@UseGuards(JwtAuthGuard)
+@UseGuards(JwtAuthGuard, RolesGuard)
 export class UserController {
     constructor(private userService: UserService) {}
 
@@ -18,33 +17,33 @@ export class UserController {
     }
 
     @Get(':id')
-    async getUser(@Param() params, @Req() req: Request): Promise<any> {
-        const user = await this.userService.findOneById(params.id);
-        return user;
-    }
+    async getUser(@Param() params, @Req() request: any): Promise<any> {
+        const user = await this.userService.findOneById(params["id"]);
 
-    @Post()
-    async createUser(@Body() userDto: CreateUserReqApiDto): Promise<any> {
-        const user = await this.userService.create(userDto);
+        if (!user) throw new BadRequestException("user with this credentials does not exist");
+        if (user.id !== request.user.sub) throw new ForbiddenException('You do not have permission to get all info about this user');
+
         return user;
     }
 
     @Patch(':id')
-    async updateUser(@Param() params: number, @Body() userDto: UpdateUserReqApiDto): Promise<any> {
-        const user = await this.userService.update(params["id"], userDto);
-        return user;
+    async updateUser(@Req() request: any, @Param() params: number, @Body() userDto: UpdateUserReqApiDto): Promise<any> {
+        const user = await this.userService.findOneById(params["id"]);
+        if (!user) throw new BadRequestException("User with this id does not exist");
+        if (request.user.role !== ERole.ADMIN && user.id !== request.user.sub) 
+            throw new ForbiddenException('You do not have permission to update this user');
+
+        const updatedUser = await this.userService.update(params["id"], userDto);
+        return updatedUser;
     }
 
     @Delete(':id')
-    async ownAccount(@Param() params: number): Promise<any> {
-        const ticket = await this.userService.remove(params["id"]);
-        return ticket;
-    }
+    async deleteUser(@Req() request: any, @Param() params: number): Promise<any> {
+        const user = await this.userService.findOneById(params["id"]);
+        if (!user) throw new BadRequestException("User with this id does not exist");
+        if (request.user.role !== ERole.ADMIN && user.id !== request.user.sub) 
+            throw new ForbiddenException('You do not have permission to delete this user');
 
-    @UseGuards(RolesGuard)
-    @RequiredRoles(ERole.ADMIN)
-    @Delete('any/:id')
-    async deleteUser(@Param() params: number): Promise<any> {
         const ticket = await this.userService.remove(params["id"]);
         return ticket;
     }
